# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for pull requests managment in teams",
        "title": "PR Reviewer Assignment Service (Test Task, Fall 2025)",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/team/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Создать команду с участниками (создает/обновляет пользователей)",
                "parameters": [
                    {
                        "description": "Данные для создания/обновления",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.AddTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Команда создана",
                        "schema": {
                            "$ref": "#/definitions/docs.AddTeamResponse"
                        }
                    },
                    "400": {
                        "description": "Команда уже существует",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Нет/неверный админский токен",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Пользователь является членом другой команды",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/team/get": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Получить команду с участниками",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальное имя команды",
                        "name": "team_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Объект команды",
                        "schema": {
                            "$ref": "#/definitions/docs.GetTeamResponse"
                        }
                    },
                    "401": {
                        "description": "Нет/неверный админский токен",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Команда не найдена",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "docs.AddTeamRequest": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/docs.TeamMember"
                    }
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "docs.AddTeamResponse": {
            "type": "object",
            "properties": {
                "team": {
                    "$ref": "#/definitions/docs.AddTeamResponseObject"
                }
            }
        },
        "docs.AddTeamResponseObject": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/docs.TeamMember"
                    }
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
...
        "docs.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "docs.GetTeamResponse": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/docs.TeamMember"
                    }
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
...
        "docs.TeamMember": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}